+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.11] - 2022-11-27
+### Added
+* table of contents in markdown output
 
 ## [0.0.11] - 2022-11-27